Move ServerData to its own class
